refactor: give the reconnect policy a module of its own
The wait between two connection attempts is decided by the devices, not by the
camera: all three read `retry_delay`, `clamp_reconnect_interval` and the
interval bounds, while `Camera` itself only clamps what it is handed. Keeping
them in the camera base module made every device import the class it is owned
by to learn how long to wait, and left the policy without a place of its own to
test.

`rosys/vision/reconnect.py` now holds the interval bounds, the jitter and the
two functions, so a device depends on the policy alone.
